Created buttons that sort entries

+++ server/public/scripts/controllers/entry.controller.js
@@ -1,20 +1,53 @@
-myApp.controller('EntryController', function ($http) {
+myApp.controller('EntryController', function ($http, $mdDialog, $mdToast) {
     const vm = this;
     vm.entryList = [];
     let timeSpent = 0;
     vm.projectList = [];
+    vm.sort = 'task';
+    vm.bool = false;
    
     vm.deleteEntry = function (entry) {
-        $http({
-            method: 'DELETE',
-            url: '/task/entry/' + entry.id
-        }).then(function (response) {
-            console.log('Entry Deleted');
-            getEntries();
-        }).catch(function (error) {
+        $mdDialog.show(
+            $mdDialog.confirm()
+                .title('Delete this task?')
+                .textContent('You cannot undo this action')
+                .ok('Delete')
+                .cancel('Cancel')
+        ).then(function() {
+            $http({
+                method: 'DELETE',
+                url: '/task/entry/' + entry.id
+            }).then(function (response) {
+                $mdToast.show(
+                    $mdToast.simple()
+                    .textContent('Task Deleted!')
+                    .hideDelay(3000)
+                );
+                console.log('Entry Deleted');
+                getEntries();
+            }).catch(function (error) {
+                $mdDialog.show(
+                    $mdDialog.alert()
+                        .title('Cannot delete a project with tasks')
+                        .ok('Okay')
+                );
+                console.log(error);
+            });//End DELETE
+        }).catch(function(error) {
             console.log(error);
-        });//End DELETE
+            $mdDialog.show(
+                $md.alert()
+                    .title('Error deleting task')
+                    .ok('Okay')
+            );
+        });
+        
     };
+
+    vm.sortFunction = function(sort) {
+        vm.sort = sort;
+        vm.bool = !vm.bool;
+    }
 
     vm.addEntry = function () {
         const start = moment.utc(vm.start, 'HH:mm');
@@ -38,7 +71,11 @@ myApp.controller('EntryController', function ($http) {
             data: vm.entryToAdd
         }).then(function(response) {
             getEntries();
-            console.log('Task created');
+            $mdToast.show(
+                $mdToast.simple()
+                .textContent('Task Added!')
+                .hideDelay(3000)
+            );
             vm.start = '';
             vm.end = '';
             vm.task = '';
@@ -58,6 +95,11 @@ myApp.controller('EntryController', function ($http) {
             console.log(response.data);
         }).catch(function (error) {
             console.log(error);
+            $mdDialog.show(
+                $mdDialog.alert()
+                    .title('Error adding task')
+                    .ok('Okay')
+            );
         });
     };
